fix(justdummies): stop a huge regex quantifier yielding a short value
RegexRepeat computed an unbounded quantifier's repetition count as
`_min + Next(0, UnboundedExtra + 1)` in int arithmetic. A minimum within
UnboundedExtra of int.MaxValue therefore wrapped negative, the repetition
loop never entered, and Generate() returned the empty string for a pattern
demanding billions of characters — 36 empty strings out of 40 draws on
`a{2147483647,}`. Silently yielding a value the pattern does not match is
the one outcome generation must never produce; every other minimum too
large to build, `a{100000,}` among them, already overran the generation
ceiling and said so.

Widen the count to long before adding the extra repetitions, so the band
`min > int.MaxValue - UnboundedExtra` walks the ceiling and reports the
overrun like any other oversized minimum. Behaviour is unchanged for every
minimum that did not wrap.

The example suite pins the regression at the coordinates it was found, plus
the four corners of the overflow band including `int.MaxValue - 8`, the last
minimum that could never wrap. The defect had an input space the examples do
not cover — the existing spread property draws its minimum from 0..4 — so
the property suite gains the half of the contract that holds for every
minimum: a value shorter than the minimum is never an outcome, whether
generation succeeds or the ceiling refuses.

Internal to the pattern walk; no public API change.

+++ JustDummies/RegexNode.cs
@@ -152,11 +152,16 @@ internal sealed class RegexRepeat : RegexNode {
 
     internal override void Append(RegexGenerationContext context) {
         if (context is null) { throw new ArgumentNullException(nameof(context)); }
-        int count = _max is int max
-                        ? context.Random.NextInt32Inclusive(_min, max)
-                        : _min + context.Random.Next(0, UnboundedExtra + 1);
+        // The unbounded count is widened to long before the extra repetitions are added: a minimum within
+        // UnboundedExtra of int.MaxValue would otherwise wrap negative, and a negative count writes nothing at all —
+        // silently yielding a value the pattern does not match, which is the one outcome generation must never
+        // produce. Widened, such a count simply walks until the generation ceiling reports the overrun, exactly as
+        // any other minimum too large to fit does.
+        long count = _max is int max
+                         ? context.Random.NextInt32Inclusive(_min, max)
+                         : (long)_min + context.Random.Next(0, UnboundedExtra + 1);
 
-        for (int repetition = 0; repetition < count; repetition++) { _child.Append(context); }
+        for (long repetition = 0; repetition < count; repetition++) { _child.Append(context); }
     }
 
 }
